# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Retrieves concordances, frequency information and collocations from language corpora",
        "title": "MQuery - query and analyze corpus data",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/concordance/{corpusId}": {
            "get": {
                "description": "Search in a corpus for concordances",
                "produces": [
                    "application/json"
                ],
                "summary": "Concordance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "An ID of a corpus to search in",
                        "name": "corpusId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The translated query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "An ID of a subcorpus",
                        "name": "subcorpus",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "json",
                            "markdown"
                        ],
                        "type": "string",
                        "default": "json",
                        "description": "For a concordance formatted in Markdown, `markdown` value can be passed",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "default": 0,
                        "description": "if 1, then markup specifying formatting and structure of text will be displayed along with tokens",
                        "name": "showMarkup",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "default": 0,
                        "description": "if 1, then text metadata (e.g. author, publication year) will be attached to each line",
                        "name": "showTextProps",
                        "in": "query"
                    },
                    {
                        "maximum": 50,
                        "minimum": 0,
                        "type": "integer",
                        "default": 10,
                        "description": "Defines number of tokens around KWIC. For a value K, the left context is floor(K / 2) and for the right context, it is ceil(K / 2).",
                        "name": "contextWidth",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/sentences/{corpusId}": {
            "get": {
                "description": "Search in a corpus for matching sentences. This is an alternative to the /concordance/{corpusId} endpoint.",
                "produces": [
                    "application/json"
                ],
                "summary": "Sentences",
                "parameters": [
                    {
                        "type": "string",
                        "description": "An ID of a corpus to search in",
                        "name": "corpusId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The translated query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "An ID of a subcorpus",
                        "name": "subcorpus",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "json",
                            "markdown"
                        ],
                        "type": "string",
                        "default": "json",
                        "description": "For a concordance formatted in Markdown, `markdown` value can be passed",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "default": 0,
                        "description": "if 1, then markup specifying formatting and structure of text will be displayed along with tokens",
                        "name": "showMarkup",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "default": 0,
                        "description": "if 1, then text metadata (e.g. author, publication year) will be attached to each line",
                        "name": "showTextProps",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/term-frequency/{corpusId}": {
            "get": {
                "description": "This endpoint retrieves the frequency, instances per million (IPM), and Average Reduced Frequency (ARF) of a searched term within a corpus. It provides a concise aggregated frequency overview for a given query, regardless of the number of concrete words (n-grams) it covers.",
                "produces": [
                    "application/json"
                ],
                "summary": "TermFrequency",
                "parameters": [
                    {
                        "type": "string",
                        "description": "An ID of a corpus to search in",
                        "name": "corpusId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The translated query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "An ID of a subcorpus",
                        "name": "subcorpus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ConcSize"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ConcSize": {
            "type": "object",
            "properties": {
                "arf": {
                    "type": "number"
                },
                "corpusSize": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "ipm": {
                    "type": "number"
                },
                "resultType": {
                    "$ref": "#/definitions/ResultType"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "ResultType": {
            "type": "string",
            "enum": [
                "conc",
                "termFrequency",
                "coll",
                "collFreqData",
                "freqs",
                "multipleFreqs",
                "corpusInfo",
                "textTypeNorms",
                "error"
            ],
            "x-enum-varnames": [
                "ResultTypeConcordance",
                "ResultTypeConcSize",
                "ResultTypeCollocations",
                "ResultTypeCollFreqData",
                "ResultTypeFreqs",
                "ResultTypeMultipleFreqs",
                "ResultTypeCorpusInfo",
                "ResultTypeTextTypeNorms",
                "ResultTypeError"
            ]
        },
...
    },
...
}